Create a new cli_do_tmux.c file, and fix the --help message.
